+++ src/array.h
@@ -204,7 +204,7 @@ size_t gc_free_all_unreferenced_arrays(void);
 void debug_dump_type_field(TYPE_FIELD t);
 void debug_dump_array(struct array *a);
 #endif
-void count_memory_in_arrays(INT32 *num_, INT32 *size_);
+void count_memory_in_arrays(size_t *num_, size_t *size_);
 PMOD_EXPORT struct array *explode_array(struct array *a, struct array *b);
 PMOD_EXPORT struct array *implode_array(struct array *a, struct array *b);
 /* Prototypes end here */
