+++ src/block_alloc_h.h
@@ -26,7 +26,7 @@ void PIKE_CONCAT3(new_,DATA,_context)(void);				\
 void PIKE_CONCAT3(really_free_,DATA,_unlocked)(struct DATA *d);		\
 void PIKE_CONCAT(really_free_,DATA)(struct DATA *d);			\
 void PIKE_CONCAT3(free_all_,DATA,_blocks)(void);			\
-void PIKE_CONCAT3(count_memory_in_,DATA,s)(INT32 *num, INT32 *size);	\
+void PIKE_CONCAT3(count_memory_in_,DATA,s)(size_t *num, size_t *size);	\
 void PIKE_CONCAT3(init_,DATA,_blocks)(void)
